common/map.c: Make a function static. common/loader.l, common/object.c, common/readable.c: Fix some format string bugs   (found using the GCC format __attribute__ for some crossfire functions).
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11548 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/readable.c
@@ -1820,7 +1820,7 @@ static char *god_info_msg(int level, char *retbuf, size_t booksize) {
     char en[BOOK_BUF];
 
     if (booksize > BOOK_BUF) {
-        LOG(llevError, "common/readable.c:god_info_msg() - passed in booksize (%d) is larger than book buffer (%d)\n", booksize, BOOK_BUF);
+        LOG(llevError, "common/readable.c:god_info_msg() - passed in booksize (%lu) is larger than book buffer (%d)\n", (unsigned long)booksize, BOOK_BUF);
         booksize = BOOK_BUF;
     }
